// frontend/src/app/core/api/models/single-question-string-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ReadSingleQuestionOptionNumberDto } from '../models/read-single-question-option-number-dto';
import { ReadSingleQuestionOptionStringDto } from '../models/read-single-question-option-string-dto';
export interface SingleQuestionStringDto {
  controlType: 'textbox' | 'dropdown' | 'number' | 'divider';
  hint?: string | null;
  key: string;
  label: string | null;
  options: (Array<(ReadSingleQuestionOptionStringDto | ReadSingleQuestionOptionNumberDto)>) | null;
  order: number;
  required: boolean;
  type: 'checkbox' | 'color' | 'date' | 'datetime-local' | 'email' | 'file' | 'hidden' | 'image' | 'month' | 'number' | 'password' | 'radio' | 'range' | 'reset' | 'search' | 'submit' | 'tel' | 'text' | 'time' | 'url' | 'week' | 'dropdown' | 'no-divider';
  value: string;
  valueType: 'string';
}